Added chart offset editing in chart editor. Press F1 while in the editor to open the online manual.

+++ Assets/Script/Chart/ChartData.cs
@@ -15,7 +15,7 @@ public class ChartData : ScriptableObject
     public float songTempo = 60f;
     public float songLength = 10f;
     public float gameplayLength = 0f;
-    public float chartOffset = 0f;
+    public int chartOffset = 0;
     public int chartJudge = 0;
     public bool isHighScoreAllowed = true;
     public bool isModifierAllowed = true;
